rename CueValue to Value
If you are using CUE from Java, it is extremely unlikely you deal
a different kin of "Value" than a CUE value. Rename CueValue to
Value to reduce the verbosity of the API.

Keep the prefix for CueContext because that appears rarely in user
code (usually once).

Also keep the prefix for CueError, as Error is likely to clash with
something.

Signed-off-by: Aram Hăvărneanu <aram@mgk.ro>