ci: run maven in CI
Install Java 22, build libcue, and test the Java API against libcue.

This currently tests the Java API against the tip of libcue. We
probably want to test against a specific version in the future.

We also make `cue-api-java` a Go module. We do this as a hack to allow
the use of `go generate`. It is not a Go module that users can use
in any way.

Signed-off-by: Aram Hăvărneanu <aram@mgk.ro>
Change-Id: I462961027146812acdf4775e4041a1c6e3b42909