gen: use envvars to specify deps' locations
This updates the "gen" script by adding a pair of environment variables
that allow developers to control where the script looks for two required
dependencies: jextract and libcue. Defaults are set up so that when the
variables aren't present the script's behaviour doesn't change,

Also: tidy up and add comments.

Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>
Change-Id: Iaabdca1496f541aa7c22315be9b8669ac51ec5fd